+++ plugin-annotation/src/main/java/org/maplibre/android/plugins/annotation/annotation/OnFillClickListener.java
@@ -1,0 +1,10 @@
+// This file is generated.
+
+package org.maplibre.android.plugins.annotation.annotation;
+
+/**
+ * Interface definition for a callback to be invoked when a fill has been clicked.
+ */
+public interface OnFillClickListener extends OnAnnotationClickListener<Fill> {
+
+}