+++ plugin-annotation/src/main/java/org/maplibre/android/plugins/annotation/annotation/Options.java
@@ -1,0 +1,6 @@
+package org.maplibre.android.plugins.annotation.annotation;
+
+public abstract class Options<T extends Annotation> {
+
+    abstract T build(long id, AnnotationManager<?, T, ?, ?, ?, ?> annotationManager);
+}